requirements.txt
Flask==0.10.1
itsdangerous==0.24
Jinja2==2.7.3
MarkupSafe==0.23
PyJWT==1.1.0
Werkzeug==0.10.4
